% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{BuildIndexFromLocalPoints}
\alias{BuildIndexFromLocalPoints}
\title{LidarIndexR -- Create spatial index by reading LAS/LAZ file headers}
\usage{
BuildIndexFromLocalPoints(
  basePath = "",
  fileType = "\\\\.las|\\\\.laz",
  folderName = "",
  pointFolder = "",
  projString = NA,
  outputCRS = NA,
  fullFileList = character(0),
  dimensionThreshold = 50000,
  headerMethod = "rcpp",
  outputFile = NULL,
  rebuild = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{basePath}{Path for a folder. Trailing slash should \emph{not}
be included.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{folderName}{Folder name on the \code{basePath} containing LAS/LAZ files.
Can be an empty string.}

\item{pointFolder}{Folder under \code{folderName} containing point files. Can be
an empty string.}

\item{projString}{A valid projection string that can be used with the \code{crs}
parameter in \code{st_sf}. \code{projString} should represent the projection
pf the point data. If using EPSG codes, do not enclose the EPSG number in quotes.}

\item{outputCRS}{A valid projection string that can be used with the \code{crs}
parameter in \code{st_transform} to reproject the index.}

\item{fullFileList}{List of point tiles. Will be used instead of generating
new list.}

\item{dimensionThreshold}{Size threshold used to omit files from the index.
This is intended to help omit invalid LAS/LAZ files from the index. If
the height or width of the point tile exceeds the threshold, the tile
will be omitted.}

\item{headerMethod}{String indicating the method that should be used to
read LAS file headers. Choices are: "direct", "lidr", or "rcpp".}

\item{outputFile}{Full path and filename on the local file system for the index
file.}

\item{rebuild}{Boolean. If TRUE, the index is always created. If FALSE,
the index is only created if it does not already exist.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}
}
\value{
An \code{sf} object (invisible) containing tile bounding boxes and attributes.
}
\description{
Reads LAS/LAZ file headers to build bounding boxes for each tile and
stores header attributes with each bounding box. Optionally writes
information to \code{outputFile}. Typically used by passing a folder
name in \code{basePath} and a file extension template such as "//.las|//.laz".
You can also specify folders under the \code{basePath} using \code{folderName}
and \code{pointFolder}. You can omit all folder information and provide a
list of point files in \code{fullFileList}. CRS information is read from
point files. If you provide \code{projString}, it will override the CRS
found in the point files.
}
\examples{
\dontrun{
BuildIndexFromPoints()
}
}
