% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{DirList}
\alias{DirList}
\title{LidarIndexR -- Retrieve a directory listing from a remote host}
\usage{
DirList(
  URL,
  fileType = NULL,
  namesOnly = FALSE,
  excludeDirectories = TRUE,
  directoryOnly = FALSE,
  dirFormat = "dir",
  ...
)
}
\arguments{
\item{URL}{URL for a folder on a remote host. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{namesOnly}{Boolean indicating you want only file names (TRUE) or all
directory information (FALSE).}

\item{excludeDirectories}{Boolean indicating you want to exclude folders
from the list of returned files/folders.}

\item{directoryOnly}{Boolean indicating you only want folder names included
in the list of returned files/folders.}

\item{dirFormat}{String indicating the expected directory format for the
\code{URL}. Valid values are "dir" if the URL returns the Date, Time,
Attribute/Size, and file name; "ls" if the URL returns directory
information similar to the UNIX \code{ls -al} command; or "" if you
want the function to try to determine the directory format. This is
ignored if the URL uses http or https schemes.}

\item{...}{Arguments passed to \code{RCurl::getURL()}. This is ignored
if the URL uses http or https schemes.}
}
\value{
A list of file names or a data frame with file names and attribute
information. The return value depends on \code{namesOnly}.
}
\description{
Retrieve a directory listing from a remote host as either a simple list
of file/folder names or a data frame with file/folder information.
}
\examples{
\dontrun{
DirList("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/Staged/")
DirList(paste0("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
   "Staged/Elevation/LPC/Projects/AK_BrooksCamp_2012/laz/"),
    "\\\\.las|\\\\.laz")
}
}
