% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{DirListByName}
\alias{DirListByName}
\title{LidarIndexR -- Retrieve a directory listing from a remote host of all files
with a specific name (various extensions)}
\usage{
DirListByName(URL, fileName, ...)
}
\arguments{
\item{URL}{URL for a folder on a remote host. Trailing slash is optional.}

\item{fileName}{A string containing the name of a single file. All files
with the same name (but different extensions) will be returned.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
A list of file names.
}
\description{
Retrieve a simple list of all files with the specified name but various
extensions.
}
\examples{
\dontrun{
DirListByName(paste0("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Hydrography/NHD/HU4/HighResolution/Shape/"),
    "NHD_H_0101_HU4_Shape.jpg")
}
}
