% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{DirListByType}
\alias{DirListByType}
\title{LidarIndexR -- Retrieve a directory listing from a remote host}
\usage{
DirListByType(URL, fileType, ...)
}
\arguments{
\item{URL}{URL for a folder on a remote host. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
A list of file names.
}
\description{
Retrieve a directory listing from a remote host as a simple list
of file names.
}
\examples{
\dontrun{
DirListByType(paste0("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Elevation/LPC/Projects/AK_BrooksCamp_2012/laz/"),
    "\\\\.las|\\\\.laz")
}
}
