% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{ExtractCRSFromLocalPoints}
\alias{ExtractCRSFromLocalPoints}
\title{LidarIndexR -- Retrieve CRS information from a LAS/LAZ point file}
\usage{
ExtractCRSFromLocalPoints(
  basePath,
  folderName,
  headerOnly = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{basePath}{Path for a folder on a remote host. Trailing slash should \emph{not}
be included.}

\item{folderName}{Folder name on the \code{basePath} containing LAS/LAZ
files.}

\item{headerOnly}{Boolean. If TRUE, only the file header is read. If FALSE,
the entire LAS/LAZ file is retrieved and then the header is read to
extract CRS information.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
String (invisible) containing a valid input value for \code{st_crs()}.
}
\description{
Reads CRS information from local LAS/LAZ files. This is done by
reading only the file header \code{headerOnly = TRUE} or retrieving the entire
file and reading the header \code{headerOnly = FALSE}.
}
\examples{
\dontrun{
FetchAndExtractCRSFromPoints()
}
}
