% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{FetchAndExtractCRSFromIndex}
\alias{FetchAndExtractCRSFromIndex}
\title{LidarIndexR -- Retrieve a shapefile (all files) and read CRS information}
\usage{
FetchAndExtractCRSFromIndex(
  baseURL,
  folderName,
  fileName,
  tempFolder,
  fetchnew = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{baseURL}{URL for a folder on a remote host. Trailing slash should \emph{not}
be included.}

\item{folderName}{Folder name on the \code{baseURL} containing the shapefile.}

\item{fileName}{A string containing the name of a single file. All
files with the same name (but different extensions) will be retrieved. The
projection file is then read to provide CRS information.}

\item{tempFolder}{A valid folder name where the files can be
temporarily stored.}

\item{fetchnew}{Boolean. If TRUE, the files are always retrieved. If FALSE,
the files are only retrieved if they do not already exist in the \code{tempFolder}.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
String containing a valid input value for \code{st_crs()}.
}
\description{
Retrieves all files related to a shapefile (same name as \code{fileName} but
different extensions) and then reads CRS information from the projection
file.
}
\examples{
\dontrun{
FetchAndExtractCRSFromIndex("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Hydrography/NHD/HU4/HighResolution/Shape/",
    "NHD_H_0101_HU4_Shape.shp",
    ".")
}
}
