% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{FetchAndExtractCRSFromPoints}
\alias{FetchAndExtractCRSFromPoints}
\title{LidarIndexR -- Retrieve CRS information from a LAS/LAZ point file}
\usage{
FetchAndExtractCRSFromPoints(
  baseURL,
  folderName,
  tempFolder,
  headerOnly = TRUE,
  fetchnew = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{baseURL}{URL for a folder on a remote host. Trailing slash should \emph{not}
be included.}

\item{folderName}{Folder name on the \code{baseURL} containing LAS/LAZ
files.}

\item{tempFolder}{A valid folder name where the LAS/LAZ file header can be
temporarily stored. The header file is deleted after use.}

\item{headerOnly}{Boolean. If TRUE, only the file header is read. If FALSE,
the entire LAS/LAZ file is retrieved and then the header is read to
extract CRS information.}

\item{fetchnew}{Boolean. If TRUE, the file is always retrieved. If FALSE,
the file is only retrieved if it does not already exist in the \code{tempFolder}.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
String (invisible) containing a valid input value for \code{st_crs()}.
}
\description{
Reads CRS information from LAS/LAZ files on a remote host. This is done by
reading only the file header \code{headerOnly = TRUE} or retrieving the entire
file and reading the header \code{headerOnly = FALSE}.
}
\examples{
\dontrun{
FetchAndExtractCRSFromPoints()
}
}
