% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{FetchAndExtractCRSFromPrj}
\alias{FetchAndExtractCRSFromPrj}
\title{LidarIndexR -- Retrieve a projection file and read CRS information}
\usage{
FetchAndExtractCRSFromPrj(
  baseURL,
  folderName,
  fileName,
  tempFolder,
  fetchnew = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{baseURL}{URL for a folder on a remote host. Trailing slash should \emph{not}
be included.}

\item{folderName}{Folder name on the \code{baseURL} containing the shapefile.}

\item{fileName}{A string containing the name of a single file. The projection
file with the same name (but extension of .prj) will be retrieved. The
projection file is then read to provide CRS information.}

\item{tempFolder}{A valid folder name where the projection file can be
temporarily stored.}

\item{fetchnew}{Boolean. If TRUE, the file is always retrieved. If FALSE,
the file is only retrieved if it does not already exist in the \code{tempFolder}.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}

\item{...}{Arguments passed to \code{RCurl::getURL()}.}
}
\value{
String (invisible) containing a valid input value for \code{st_crs()}.
}
\description{
Retrieves only the projection file related to a shapefile (same name as
\code{fileName} but with .prj extensions) and then reads CRS information
from the projection file.
}
\examples{
\dontrun{
FetchAndExtractCRSFromPrj("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Hydrography/NHD/HU4/HighResolution/Shape/",
    "NHD_H_0101_HU4_Shape.prj",
    ".")
}
}
