% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{LocalDirList}
\alias{LocalDirList}
\title{LidarIndexR -- Retrieve a directory listing}
\usage{
LocalDirList(
  path,
  fileType = NULL,
  namesOnly = TRUE,
  directoryOnly = FALSE,
  ...
)
}
\arguments{
\item{path}{Path for a folder. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{namesOnly}{Boolean indicating you want only file names (TRUE) or all
directory information (FALSE).}

\item{directoryOnly}{Boolean indicating you only want folder names included
in the list of returned files/folders.}

\item{...}{Arguments passed to \code{RCurl::getURL()}. This is ignored
if the path uses http or https schemes.}
}
\value{
A list of file names (invisible) or a data frame (invisible) with file names and attribute
information. The return value depends on \code{namesOnly}.
}
\description{
Retrieve a directory listing from a local file system as either a simple list
of file/folder names or a data frame with file/folder information.
}
\examples{
\dontrun{
DirList("vdelivery/Datasets/Staged/")
}
}
