% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{LocalDirListByType}
\alias{LocalDirListByType}
\title{LidarIndexR -- Retrieve a directory listing from a remote host}
\usage{
LocalDirListByType(URL, fileType, ...)
}
\arguments{
\item{URL}{URL for a folder on a remote host. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{...}{Arguments passed to \code{RCurl::getURL()}. This is ignored
if the URL uses http or https schemes.}
}
\value{
A list of file names.
}
\description{
Retrieve a directory listing from a remote host as a simple list
of file names.
}
\details{
DirListByType is an alias for \code{DirList(URL, fileType = fileType, namesOnly = TRUE, ...)}
}
\examples{
\dontrun{
DirListByType(paste0("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Elevation/LPC/Projects/AK_BrooksCamp_2012/laz/"),
    "\\\\.las|\\\\.laz")
}
}
