% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{makeFileSize}
\alias{makeFileSize}
\title{LidarIndexR -- Utility function to convert file size string to actual
file size.}
\usage{
makeFileSize(sizeString)
}
\arguments{
\item{sizeString}{A string containing the file size or "-" for directories.}
}
\value{
A (invisible) double value for files and a string for directories.
}
\description{
This is a helper function used in the LidarIndexR package to help read
file sizes from http or https directory listings. It looks for a trailing
character to determine the multipliuer for the numeric portion of the size
value. In the case of directories, the \code{sizeString} will be "-". This
value is explicitly recognized.
}
