% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadLASHeader}
\alias{ReadLASHeader}
\title{LidarIndexR -- Read the header for a local LAS/LAZ file}
\usage{
ReadLASHeader(path)
}
\arguments{
\item{path}{Path for a LAS/LAZ file.}
}
\value{
A data frame containing the CRS WKT information string and
header info. If file has no CRS information, the crs column in the data frame
will be an empty string.
}
\description{
Read the header information for a LAS/LAZ file including CRS information.
Only the file header is read (including VLRs) so you don't have to worry
about the size of the LAS/LAZ file. This function is implemented in C++
and is 2+ orders of magnitude faster than \code{ReadLocalLASHeader()}.
}
\examples{
\dontrun{
ReadLocalLASHeader()
}
}
