% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{ReadLocalLASHeader}
\alias{ReadLocalLASHeader}
\title{LidarIndexR -- Read the header for a local LAS/LAZ file}
\usage{
ReadLocalLASHeader(path, quiet = FALSE)
}
\arguments{
\item{path}{Path for a LAS/LAZ file.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}
}
\value{
A data frame (invisible) containing header information.
}
\description{
Reads the header of a LAS/LAZ file and returns header information in a
data frame. Only the header is read.
}
\examples{
\dontrun{
ReadLocalLASHeader()
}
}
