% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{ReadLocalLASInfo}
\alias{ReadLocalLASInfo}
\title{LidarIndexR -- Read LAS/LAZ header info and CRS information}
\usage{
ReadLocalLASInfo(path, quiet = TRUE)
}
\arguments{
\item{path}{Path for a LAS/LAZ file.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}
}
\value{
A list (invisible) containing the CRS WKT information string and a
data frame with header info. If file has no CRS information, the first element
in the list is an empty string.
}
\description{
Read the header information for a LAS/LAZ file including CRS information.
Only the file header is read (including VLRs) so you don't have to worry
about the size of the LAS/LAZ file.
}
\examples{
\dontrun{
ReadLocalLASInfo()
}
}
