% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{ReadRemoteLASProjection}
\alias{ReadRemoteLASProjection}
\title{LidarIndexR -- Read LAS/LAZ header and extract CRS information}
\usage{
ReadRemoteLASProjection(URL, tempFolder, quiet = FALSE)
}
\arguments{
\item{URL}{URL for a LAS/LAZ file on a remote host.}

\item{tempFolder}{A valid folder name where the LAS/LAZ file header can be
temporarily stored. The header file is deleted after use.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}
}
\value{
A string containing the CRS information retrieved from the LAS/LAZ
header.
}
\description{
Download the header for a LAS/LAZ file and read CRS information from the header.
Only the file header is downloaded (including VLRs) so you don't have to worry
about the size of the LAS/LAZ file.
}
\examples{
\dontrun{
ReadRemoteLASProjection()
}
}
