% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{unsignedFourByteIntToDouble}
\alias{unsignedFourByteIntToDouble}
\title{LidarIndexR -- Utility function to help read unsigned long data types}
\usage{
unsignedFourByteIntToDouble(i)
}
\arguments{
\item{i}{A 4-byte block of data.}
}
\value{
An (invisible) double value.
}
\description{
This is a helper function used in the LidarIndexR package to help read
\code{unsigned long} data types. It takes a 4-byte block and interprets
it as and \code{unsigned long} value. The function is used mainly to read
value from LAS file headers.
}
