% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{BuildProjectPolygonFromIndex}
\alias{BuildProjectPolygonFromIndex}
\title{LidarIndexR -- Create polygon(s) representing the area covered by a tile index}
\usage{
BuildProjectPolygonFromIndex(
  indexFile,
  projectIdentifier,
  nx = 512,
  ny = 512,
  outputCRS = NULL,
  appendInfo = NULL,
  outputFile = NULL,
  rebuild = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{indexFile}{Full path and filename on the local file system for the tile index
file.}

\item{projectIdentifier}{Character string containing an identifier for the project
area. This will be added to each feature in the return \code{sf} object.}

\item{nx}{Integer specifying the number of cells (width) of the raster used
to merge polygons.}

\item{ny}{Integer specifying the number of cells (height) of the raster used
to merge polygons.}

\item{outputCRS}{A valid projection string that can be used with the \code{crs}
parameter in \code{st_transform} to reproject the index. If using EPSG codes,
do not enclose the EPSG number in quotes.}

\item{appendInfo}{Data frame (single row) with values to append to each feature
in the return \code{sf} object. If NULL, no information is appended.
If \code{appendInfo} is \code{sf} object, the geometry will be removed
before appending to each feature.}

\item{outputFile}{Full path and file name on the local file system for the index
file.}

\item{rebuild}{Boolean. If TRUE, the index is always created. If FALSE,
the index is only created if it does not already exist.}

\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}
}
\value{
An \code{sf} object containing the polygon(s) for the project area.
}
\description{
Creates a bounding polygon(s) given a set of point tile polygons. This is useful to
create project polygons to display the areas covered by a lidar project.
}
\examples{
\dontrun{
BuildProjectPolygonFromIndex()
}
}
