% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarIndexR.R
\name{DirList_http}
\alias{DirList_http}
\title{LidarIndexR -- Retrieve a directory listing from a remote http(s) host}
\usage{
DirList_http(
  URL,
  fileType = NULL,
  namesOnly = FALSE,
  excludeDirectories = TRUE,
  directoryOnly = FALSE
)
}
\arguments{
\item{URL}{URL for a folder on a remote host. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{namesOnly}{Boolean indicating you want only file names (TRUE) or all
directory information (FALSE).}

\item{excludeDirectories}{Boolean indicating you want to exclude folders
from the list of returned files/folders.}

\item{directoryOnly}{Boolean indicating you only want folder names included
in the list of returned files/folders.}
}
\value{
A list of file names or a data frame with file names and attribute
information. The return value depends on \code{namesOnly}.
}
\description{
Retrieve a http or https directory listing from a remote host as either
a simple list of file/folder names or a data frame with file/folder information.
}
\examples{
\dontrun{
DirList_http(pasteo("https://rockyweb.usgs.gov/vdelivery/Datasets/Staged/",
  "Elevation/LPC/Projects/AK_BrooksCamp_2012/laz/"))
}
}
