% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{LocalDirListByType}
\alias{LocalDirListByType}
\title{LidarIndexR -- Retrieve a directory listing}
\usage{
LocalDirListByType(path, fileType, ...)
}
\arguments{
\item{path}{Path for a folder on a remote host. Trailing slash is optional.}

\item{fileType}{Any valid string for the pattern parameter in \code{grep()}.
"$" will be appended to the string to search for file/folder names ending
with values in \code{fileType}.}

\item{...}{Arguments passed to \code{RCurl::getURL()}. This is ignored
if the path uses http or https schemes.}
}
\value{
A list of file names (invisible).
}
\description{
Retrieve a directory listing from a local file system as a simple list
of file names.
}
\details{
DirListByType is an alias for \code{DirList(path, fileType = fileType, namesOnly = TRUE, ...)}
}
\examples{
\dontrun{
DirListByType(paste0("ftp://rockyftp.cr.usgs.gov/vdelivery/Datasets/",
    "Staged/Elevation/LPC/Projects/AK_BrooksCamp_2012/laz/"),
    "\\\\.las|\\\\.laz")
}
}
