% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalFileFunctions.R
\name{ReadLocalLASProjection}
\alias{ReadLocalLASProjection}
\title{LidarIndexR -- Read LAS/LAZ header and extract CRS information}
\usage{
ReadLocalLASProjection(url, quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean to control display of status information. If TRUE,
information is \emph{not} displayed. Otherwise, status information is displayed.}

\item{URL}{URL for a LAS/LAZ file on a remote host.}
}
\value{
A string containing the CRS information retrieved from the LAS/LAZ
header.
}
\description{
Download the header for a LAS/LAZ file and read CRS information from the header.
Only the file header is downloaded (including VLRs) so you don't have to worry
about the size of the LAS/LAZ file.
}
\examples{
\dontrun{
ReadLocalLASProjection()
}
}
